tests/mediacapture-streams/GUM-deny.https.html

<!doctype html>
<html>
<head>
  <title>getUserMedia() triggers error callback when auth is denied</title>
  <link rel="author" title="Dr. A. Gouaillard" href="mailto:agouaillard@gmail.com"/>
  <link rel="help" href="http://w3c.github.io/mediacapture-main/getusermedia.html#methods-5">
  <link rel="help" href="http://w3c.github.io/mediacapture-main/getusermedia.html#navigatorusermediaerrorcallback">
</head>
<body>
  <p class="instructions">When prompted, <strong>please deny</strong> access to
    the video stream.</p>
  <h1 class="instructions">Description</h1>
  <p class="instructions">This test checks that the error callback is triggered
    when user denies access to the video stream.</p>
  <div id='log'></div>
  <script src="../resources/testharness.js"></script>
  <script src="../resources/testharnessreport.js"></script>
  <script>

function make_script() {
  return '<script>' +
         '  let mediastream;' +
         '  window.onmessage = message => {' +
         '    if (message.data === "INIT") {' +
         '      navigator.mediaDevices.getUserMedia({video: true})' +
         '      .then((media) => {mediastream = media;parent.postMessage( {result: mediastream.active} , "*")})' +
         '    } else if (message.data === "ACQUIRED") {' +
         '      parent.postMessage( {result: mediastream.active} , "*");' +
         '    }' +
         '  };' +
         '<\/script>';
}

function load_iframe(src) {
  return new Promise(resolve => {
    const iframe = document.createElement("iframe");
    iframe.onload = () => { resolve(iframe); };
    iframe.srcdoc = src;
    iframe.style.display = "none";
    document.documentElement.appendChild(iframe);
  });
}

function wait_for_message(iframe) {
  return new Promise(resolve => {
    self.addEventListener("message", function listener(e) {
      if (e.source === iframe.contentWindow) {
        resolve(e.data);
        self.removeEventListener("message", listener);
      }
    });
  });
}

promise_test(t => {
  let iframe;
  return load_iframe(make_script())
  .then(ifr => {
    iframe = ifr;
    iframe.contentWindow.postMessage('INIT', '*');
    return wait_for_message(iframe);
  })
  .then(message => {
    console.log(message);
    iframe.contentWindow.postMessage('ACQUIRED', '*');
    return wait_for_message(iframe);
  })
  .then(message => {
    console.log(message);
  });
}, "Test that the WakeLock active is independent.");
  </script>
</body>
</html>


<!DOCTYPE html>
<meta charset="utf-8">
<title>WakeLock independent Test</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://w3c.github.io/wake-lock/">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>

<script>

function make_script() {
  return '<script>' +
         '  let iframeWakeLock, iframeRequest;' +
         '  window.onmessage = message => {' +
         '    if (message.data === "INIT") {' +
         '      navigator.getWakeLock("screen")' +
         '      .then(wakeLock => {' +
         '        iframeWakeLock = wakeLock;' +
         '        parent.postMessage(iframeWakeLock.active, "*");' +
         '      });' +
         '    } else if (message.data === "ACQUIRED") {' +
         '      iframeRequest = iframeWakeLock.createRequest();' +
         '      parent.postMessage(iframeWakeLock.active, "*");' +
         '    } else if (message.data === "RELEASED") {' +
         '      iframeRequest.cancel();' +
         '      parent.postMessage(iframeWakeLock.active, "*");' +
         '    }' +
         '  };' +
         '<\/script>';
}

function load_iframe(src) {
  return new Promise(resolve => {
    const iframe = document.createElement("iframe");
    iframe.onload = () => { resolve(iframe); };
    iframe.srcdoc = src;
    iframe.style.display = "none";
    document.documentElement.appendChild(iframe);
  });
}

function wait_for_message(iframe) {
  return new Promise(resolve => {
    self.addEventListener("message", function listener(e) {
      if (e.source === iframe.contentWindow) {
        resolve(e.data);
        self.removeEventListener("message", listener);
      }
    });
  });
}

promise_test(async t => {
  let iframe;
  return load_iframe(make_script())
  .then(ifr => {
    iframe = ifr;
    iframe.contentWindow.postMessage('INIT', '*');
    return wait_for_message(iframe);
  })
  .then(message => {
    console.log(message);
    iframe.contentWindow.postMessage('ACQUIRED', '*');
    return wait_for_message(iframe);
  })
  .then(message => {
    console.log(message);
  });

  let iframe = document.createElement('iframe');
  iframe.srcdoc = '<script>' +
         '  let iframeWakeLock, iframeRequest;' +
         '  window.onmessage = message => {' +
         '    if (message.data === "INIT") {' +
         '      navigator.getWakeLock("screen")' +
         '      .then(wakeLock => {' +
         '        iframeWakeLock = wakeLock;' +
         '        parent.postMessage(iframeWakeLock.active, "*");' +
         '      })' +
         '    } else if (message.data === "ACQUIRED") {' +
         '      iframeRequest = iframeWakeLock.createRequest();' +
         '      parent.postMessage(iframeWakeLock.active, "*");' +
         '    } else if (message.data === "RELEASED") {' +
         '      iframeRequest.cancel();' +
         '      parent.postMessage(iframeWakeLock.active, "*");' +
         '    }' +
         '  };' +
         '<\/script>';
  
  let wakeLock1 = await navigator.getWakeLock("screen");
  let request1 = wakeLock1.createRequest();
  let wakeLock2 = await navigator.getWakeLock("screen");
  iframe.onload = () => iframe.contentWindow.postMessage('INIT', '*');
  document.body.appendChild(iframe);
  let initMessage = await window.onmessage;
  assert_true(wakeLock1.active, "the active is true if the wake lock has been acquired");
  assert_false(wakeLock2.active, "the active is false before the wake lock is acquired");
  assert_false(initMessage, "the active is false before the iframe wake lock is acquired");

  iframe.contentWindow.postMessage('ACQUIRED', '*');
  let acquiredMessage = await window.onmessage;
  assert_true(wakeLock1.active, "the active is true if the wake lock has been acquired");
  assert_false(wakeLock2.active, "the active is false before the wake lock is acquired");
  assert_true(acquiredMessage, "the active is true if the iframe wake lock has been acquired");

  request1.cancel();
  let request2 = wakeLock2.createRequest();
  iframe.contentWindow.postMessage('RELEASED', '*');
  let releasedMessage = await window.onmessage;
  assert_false(wakeLock1.active, "the active is false if the wake lock has been released");
  assert_true(wakeLock2.active, "the active is true if the wake lock has been acquired");
  assert_false(acquiredMessage, "the active is false if the iframe wake lock has been released");
}, "Test that the WakeLock active is independent.");

</script>
